requirements.txt
# Usage: pip install -r requirements.txt

# Base ----------------------------------------
tensorflow==2.6.0
keras==2.6

# Logging -------------------------------------
# tensorboard>=2.4.1
# wandb

# Plotting ------------------------------------
seaborn==0.11.2    # a Python data visualization library based on matplotlib

# Export --------------------------------------
scikit_learn==1.0.2  # CoreML quantization

# coremltools>=4.1  # CoreML export
# onnx>=1.9.0  # ONNX export
# onnx-simplifier>=0.3.6  # ONNX simplifier
# tensorflow>=2.4.1  # TFLite export
# tensorflowjs>=3.9.0  # TF.js export
# openvino-dev  # OpenVINO export

# Extras --------------------------------------
matplotlib==3.5.1
PyYAML==6.0
ipykernel==6.15.1
six==1.15.0 # a Python 2 and 3 compatibility library
autopep8==1.7.0 # automatically formats Python code (Usage: Shift + Alt + F)
split-folders==0.5.1 # Split folders with files (e.g. images) into train, validation and test (dataset) folders.
opencv-python==4.6.0.66
scikit-image==0.19.3       # Image processing in Python
xlwings==0.27.15


# ipython  # interactive notebook
# psutil  # system utilization
# thop  # FLOPs computation
# albumentations>=1.0.3
# pycocotools>=2.0  # COCO mAP
# roboflow
